Added sparkline component
