+++ src/components/data-viz/Sparkline/index.js
@@ -1,0 +1,1 @@
+export default from './Sparkline'
